requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    make upgrade
#
analytics-python==1.4.0
    # via -r requirements/base.txt
asgiref==3.4.1
    # via
    #   -r requirements/base.txt
    #   django
astroid==2.9.3
    # via
    #   pylint
    #   pylint-celery
attrs==21.4.0
    # via pytest
backoff==1.10.0
    # via
    #   -r requirements/base.txt
    #   analytics-python
certifi==2021.10.8
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.15.0
    # via
    #   -r requirements/base.txt
    #   cryptography
charset-normalizer==2.0.10
    # via
    #   -r requirements/base.txt
    #   requests
click==8.0.3
    # via
    #   click-log
    #   code-annotations
    #   edx-lint
click-log==0.3.2
    # via edx-lint
code-annotations==1.2.0
    # via
    #   -r requirements/test.in
    #   edx-lint
coreapi==2.3.3
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via
    #   -r requirements/base.txt
    #   coreapi
coverage[toml]==6.2
    # via
    #   -r requirements/test.in
    #   pytest-cov
cryptography==36.0.1
    # via
    #   -r requirements/base.txt
    #   pyjwt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   -r requirements/base.txt
    #   python3-openid
    #   social-auth-core
distlib==0.3.4
    # via virtualenv
    # via
    #   -r requirements/base.txt
    #   django-cors-headers
    #   django-crum
    #   django-extensions
    #   djangorestframework
    #   drf-jwt
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
django-cors-headers==3.11.0
    # via -r requirements/base.txt
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-dynamic-fixture==3.1.2
    # via -r requirements/test.in
django-extensions==3.1.5
    # via -r requirements/base.txt
django-rest-swagger==2.2.0
    # via -r requirements/base.txt
django-waffle==2.3.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.13.1
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
    #   drf-jwt
    #   edx-drf-extensions
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-auth-backends==4.0.1
    # via -r requirements/base.txt
edx-django-release-util==1.1.1
    # via -r requirements/base.txt
edx-django-utils==4.4.1
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   edx-rest-api-client
edx-drf-extensions==8.0.1
    # via -r requirements/base.txt
edx-lint==5.2.1
    # via -r requirements/test.in
edx-opaque-keys==2.2.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-rest-api-client==5.4.1
    # via -r requirements/base.txt
filelock==3.4.2
    # via
    #   tox
    #   virtualenv
future==0.18.2
    # via
    #   -r requirements/base.txt
    #   pyjwkest
idna==3.3
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via pylint
itypes==1.2.0
    # via
    #   -r requirements/base.txt
    #   coreapi
jinja2==3.0.3
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   coreschema
lazy-object-proxy==1.7.1
    # via astroid
markupsafe==2.0.1
    # via
    #   -r requirements/base.txt
    #   jinja2
mccabe==0.6.1
    # via pylint
mock==4.0.3
    # via -r requirements/test.in
monotonic==1.6
    # via
    #   -r requirements/base.txt
    #   analytics-python
mysqlclient==2.1.0
    # via -r requirements/base.txt
newrelic==7.2.4.171
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
oauthlib==3.1.1
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
openapi-codec==1.3.2
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
packaging==21.3
    # via
    #   pytest
    #   tox
pbr==5.8.0
    # via
    #   -r requirements/base.txt
    #   stevedore
platformdirs==2.4.1
    # via
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via
    #   pytest
    #   tox
psutil==5.9.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
py==1.11.0
    # via
    #   pytest
    #   tox
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pycryptodomex==3.12.0
    # via
    #   -r requirements/base.txt
    #   pyjwkest
pyjwkest==1.4.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
pyjwt[crypto]==2.3.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-auth-backends
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   social-auth-core
pylint==2.12.2
    # via
    #   edx-lint
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via edx-lint
pylint-django==2.5.0
    # via edx-lint
pylint-plugin-utils==0.7
    # via
    #   pylint-celery
    #   pylint-django
pymongo==4.0.1
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pyparsing==3.0.6
    # via packaging
pytest==6.2.5
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==3.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   analytics-python
    #   edx-drf-extensions
python-slugify==5.0.2
    # via code-annotations
python3-openid==3.2.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
pytz==2021.3
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-release-util
requests==2.27.1
    # via
    #   -r requirements/base.txt
    #   analytics-python
    #   coreapi
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   pyjwkest
    #   requests-oauthlib
    #   slumber
    #   social-auth-core
requests-oauthlib==1.3.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
semantic-version==2.8.5
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
simplejson==3.17.6
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   analytics-python
    #   django-dynamic-fixture
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-drf-extensions
    #   edx-lint
    #   pyjwkest
    #   python-dateutil
    #   tox
    #   virtualenv
slumber==0.7.1
    # via
    #   -r requirements/base.txt
    #   edx-rest-api-client
social-auth-app-django==5.0.0
    # via
    #   -r requirements/base.txt
    #   edx-auth-backends
social-auth-core==4.1.0
    # via
    #   -r requirements/base.txt
    #   edx-auth-backends
    #   social-auth-app-django
sqlparse==0.4.2
    # via
    #   -r requirements/base.txt
    #   django
stevedore==3.5.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via
    #   pylint
    #   pytest
    #   tox
tomli==2.0.0
    # via coverage
tox==3.24.5
    # via -r requirements/test.in
typing-extensions==4.0.1
    # via
    #   astroid
    #   pylint
uritemplate==4.1.1
    # via
    #   -r requirements/base.txt
    #   coreapi
urllib3==1.26.8
    # via
    #   -r requirements/base.txt
    #   requests
virtualenv==20.13.0
    # via tox
wrapt==1.13.3
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
